html/editing/dnd/the-datatransfer-interface/DataTransferItem_drag_string.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>HTML Test: DataTransferItem_drag_string</title>
    <link rel="author" title="Intel" href="http://www.intel.com">
    <link rel="help" href="http://www.w3.org/TR/2012/CR-html5-20121217/editing.html#dom-datatransferitem-getasstring">
    <meta name="flags" content="interact">
    <meta name="assert" content="Check the DataTransferItemList interface when dragging a string">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <h2>Steps:</h2>
    <ol>
      <li>Select all characters in the first textbox below.</li>
      <li>Drag it to the second textbox.</li>
      <li>Drop it.</li>
      <li>Check the result.</li>
    </ol>
    <h2>Expected result:</h2>
    <p>PASS</p>
    <input id="test1" type="text" value="Drag me">
    <input id="test2" type="text">
    <div id="log"></div>
    <script>
      setup({explicit_done: true, explicit_timeout: true});

      on_event(document.getElementById("test1"), "dragstart", function(evt) {
        test(function () {
          var item = evt.dataTransfer.items[0];
          var file1 = item.getAsFile();
          assert_equals(file1, null, "The value is null when a string calls the getAsFile method");
        }, "Check if the value is null when a string calls the getAsFile method");

        test(function () {
          var item = evt.dataTransfer.items[0];
          item.getAsString( function(data) {
            assert_equals(data, "Drag me", "The getAsString method should return the dragged string");
          });
        }, "Check if the getAsString method returns the dragged string");

        test(function ()  {
          assert_equals(evt.dataTransfer.items[0].kind, "string", "The kind attribute should be 'text/plain'.");
        }, "Check the kind attribute when dragging a string");

        test(function () {
          assert_equals(evt.dataTransfer.items[0].type, "text/plain", "The type attribute should be 'string'.");
        }, "Check the type attribute when dragging a string");

        test(function () {
          var item = evt.dataTransfer.items[0];
          item.kind = "Files";
          item.getAsString( function(data) {
            assert_unreached("The callback should not be invoked when the kind of DataTransferItem is not 'string'.");
          });
        }, "Check if the callback can not be invoked when the kind of DataTransferItem is not 'string'.");
      });

      on_event(document.getElementById("test2"), "drop", function(evt) {
        test(function () {
          var item = evt.dataTransfer.items[0];
          item.getAsString( function(data) {
            assert_equals(data, "Drag me", "The getAsString method should return the dragged string");
          });
        }, "Check if the callback of getAsString method can be invoked in readonly mode");
        done();
      });

      on_event(document.getElementById("test2"), "dragenter", function(evt) {
        test(function () {
          var item = evt.dataTransfer.items[0];
          item.getAsString( function(data) {
            assert_unreached("The callback should not be invoked.");
          });
        }, "Check if the callback of getAsString method can not be invoked in read/write or readonly mode");
      });
    </script>
  </body>
</html>
